Tilemap for noHang Background
I'd like to have some sort of background, where if the hanger anchor lands on it, it won't stick. I added a placeholder tile and layer for this purpose.  Right now, all that works is the fact that the player is painted in front of this layer.

What doesn't work:
* The tilemap_base sprites will probably be painted over by these, which isn't desired, so layers need more work
* Not sure how to get the hanger to check where it lands, and not hold position if it lands on a noHang zone