Correct Grapple on Moving, broken on BG
I tried to get the hinge, hook spirte, and line, to behavie properly on a moving platform.  This seems to work.

Unfortunately, the line and hook no longer work for simple background attachments.  The Hinge still seems to work correctly, but the sprites seem to get sent near the origin.  Very likely just a coordinate axis transform issue.