Multiple Forms to choose from, Toggling, and UI shows current selection
Adding new forms right now is as simple as adding another FormHolder_ child to the player, and choosing the sprite you want.

Morphing Manager gets a list of all of those children, and lets you use chosen form up and down buttons to toggle through them (Q and E for right now).

When you hit Q or E, you move up or down the list of available forms (skipping the base, which is always what you revert back to when you release left shift).

I added a UI overlay, which right now only shows which form you'll turn into when you press shift.  If you're already morphed, you can't morph into that form until you revert to base first.  Hitting Q or E updates this UI icon, as well.

All these inputs are handled by the MorphingManager in Update.  We may want to move it all to CharacterController so that it's in one place.  That wouldn't be hard, but I'm leaving it as it is for now.

As of right now, it's still set up so that changing form only changes the sprite.  No other properties are changed yet.

Also, There's a bug in grappling that I first noticed here, but it probably goes way back.  If you mouse click on a valid grapple location, but the max distance clamp ends up putting the grapple in an invalid location, you hang from the invalid location.  This is probably because there's no validity clamp on the check.  The reverse is also a bug.  If you click on an invalid location, but the max distance clamp lands on a valid one, you don't hang.