Added an "Oooompf" box.
Created a little script that applies a vertical impulse to the player when the player hits it. I tried doing this with an effector, but I couldn't customize the applied force enough to do what I want, or make it an impulse for that matter.

In the inspector, you can set the final velocity you want the oomfbox to try to give.

If the player hits the box with any negative vertical velocity, the impulse applied to the player is strong enough to counteract this initial negative momentum, AND result in the desired final upward velocity.

If the player initially has positive vertical momentum, I didn't want to take that away, so the impulse applied just adds to that velocity.

I'd like to generalize this so that I can point these boxes in any direction, but that will probably take some finagling with vectors...maybe.  Maybe just rotating the box and converting the local up to world would give me the vector I want.